+++ zcash_primitives/Cargo.toml
@@ -47,6 +47,7 @@ rand_xorshift = "0.2"
 [features]
 transparent-inputs = ["ripemd160", "secp256k1"]
 test-dependencies = ["proptest"]
+zfuture = []
 
 [[bench]]
 name = "note_decryption"
