+++ zcash_primitives/src/transaction/components.rs
@@ -5,13 +5,21 @@ use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
 use ff::PrimeField;
 use group::GroupEncoding;
 
-use std::convert::TryFrom;
 use std::io::{self, Read, Write};
 
-use crate::extensions::transparent as tze;
-use crate::legacy::Script;
-use crate::redjubjub::{PublicKey, Signature};
-use crate::serialize::{CompactSize, Vector};
+#[cfg(feature = "zfuture")]
+use std::convert::TryFrom;
+
+use crate::{
+    legacy::Script,
+    redjubjub::{PublicKey, Signature},
+};
+
+#[cfg(feature = "zfuture")]
+use crate::{
+    extensions::transparent as tze,
+    serialize::{CompactSize, Vector},
+};
 
 pub mod amount;
 pub use self::amount::Amount;
@@ -121,19 +129,22 @@ impl TxOut {
     }
 }
 
+#[cfg(feature = "zfuture")]
+fn to_io_error(_: std::num::TryFromIntError) -> io::Error {
+    io::Error::new(io::ErrorKind::InvalidData, "value out of range")
+}
+
 #[derive(Clone, Debug, PartialEq)]
+#[cfg(feature = "zfuture")]
 pub struct TzeIn {
     pub prevout: OutPoint,
     pub witness: tze::Witness,
 }
 
-fn to_io_error(_: std::num::TryFromIntError) -> io::Error {
-    io::Error::new(io::ErrorKind::InvalidData, "value out of range")
-}
-
 /// Transaction encoding and decoding functions conforming to ZIP-222
 ///
 /// https://zips.z.cash/zip-0222#encoding-in-transactions
+#[cfg(feature = "zfuture")]
 impl TzeIn {
     /// Convenience constructor
     pub fn new(prevout: OutPoint, extension_id: u32, mode: u32) -> Self {
@@ -198,11 +209,13 @@ impl TzeIn {
 }
 
 #[derive(Clone, Debug, PartialEq)]
+#[cfg(feature = "zfuture")]
 pub struct TzeOut {
     pub value: Amount,
     pub precondition: tze::Precondition,
 }
 
+#[cfg(feature = "zfuture")]
 impl TzeOut {
     pub fn read<R: Read>(mut reader: &mut R) -> io::Result<Self> {
         let value = {
