+++ zcash_primitives/src/transaction/sighash.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "zfuture")]
 use std::convert::TryInto;
 
 use blake2b_simd::{Hash as Blake2bHash, Params as Blake2bParams};
@@ -5,19 +6,24 @@ use byteorder::{LittleEndian, WriteBytesExt};
 use ff::PrimeField;
 use group::GroupEncoding;
 
+use crate::{consensus, legacy::Script};
+
+#[cfg(feature = "zfuture")]
 use crate::{
-    consensus,
     extensions::transparent::Precondition,
-    legacy::Script,
     serialize::{CompactSize, Vector},
 };
 
 use super::{
-    components::{
-        Amount, JSDescription, OutputDescription, SpendDescription, TxIn, TxOut, TzeIn, TzeOut,
-    },
+    components::{Amount, JSDescription, OutputDescription, SpendDescription, TxIn, TxOut},
     Transaction, TransactionData, OVERWINTER_VERSION_GROUP_ID, SAPLING_TX_VERSION,
-    SAPLING_VERSION_GROUP_ID, ZFUTURE_VERSION_GROUP_ID,
+    SAPLING_VERSION_GROUP_ID,
+};
+
+#[cfg(feature = "zfuture")]
+use super::{
+    components::{TzeIn, TzeOut},
+    ZFUTURE_VERSION_GROUP_ID,
 };
 
 const ZCASH_SIGHASH_PERSONALIZATION_PREFIX: &[u8; 12] = b"ZcashSigHash";
@@ -27,10 +33,15 @@ const ZCASH_OUTPUTS_HASH_PERSONALIZATION: &[u8; 16] = b"ZcashOutputsHash";
 const ZCASH_JOINSPLITS_HASH_PERSONALIZATION: &[u8; 16] = b"ZcashJSplitsHash";
 const ZCASH_SHIELDED_SPENDS_HASH_PERSONALIZATION: &[u8; 16] = b"ZcashSSpendsHash";
 const ZCASH_SHIELDED_OUTPUTS_HASH_PERSONALIZATION: &[u8; 16] = b"ZcashSOutputHash";
+
+#[cfg(feature = "zfuture")]
 const ZCASH_TZE_INPUTS_HASH_PERSONALIZATION: &[u8; 16] = b"Zcash_TzeInsHash";
+#[cfg(feature = "zfuture")]
 const ZCASH_TZE_OUTPUTS_HASH_PERSONALIZATION: &[u8; 16] = b"ZcashTzeOutsHash";
 
+#[cfg(feature = "zfuture")]
 const ZCASH_TZE_SIGNED_INPUT_TAG: &[u8; 1] = &[0x00];
+#[cfg(feature = "zfuture")]
 const ZCASH_TRANSPARENT_SIGNED_INPUT_TAG: &[u8; 1] = &[0x01];
 
 pub const SIGHASH_ALL: u32 = 1;
@@ -56,11 +67,14 @@ macro_rules! update_hash {
     };
 }
 
+/// This is a private enum; when `cfg(feature = "zfuture")` is not
+/// enabled, SigHashVersion::ZFuture is not constructable.
 #[derive(PartialEq)]
 enum SigHashVersion {
     Sprout,
     Overwinter,
     Sapling,
+    #[cfg(feature = "zfuture")]
     ZFuture,
 }
 
@@ -70,6 +84,7 @@ impl SigHashVersion {
             match tx.version_group_id {
                 OVERWINTER_VERSION_GROUP_ID => SigHashVersion::Overwinter,
                 SAPLING_VERSION_GROUP_ID => SigHashVersion::Sapling,
+                #[cfg(feature = "zfuture")]
                 ZFUTURE_VERSION_GROUP_ID => SigHashVersion::ZFuture,
                 _ => unimplemented!(),
             }
@@ -172,6 +187,7 @@ fn shielded_outputs_hash(shielded_outputs: &[OutputDescription]) -> Blake2bHash 
         .hash(&data)
 }
 
+#[cfg(feature = "zfuture")]
 fn tze_inputs_hash(tze_inputs: &[TzeIn]) -> Blake2bHash {
     let mut data = vec![];
     for tzein in tze_inputs {
@@ -183,6 +199,7 @@ fn tze_inputs_hash(tze_inputs: &[TzeIn]) -> Blake2bHash {
         .hash(&data)
 }
 
+#[cfg(feature = "zfuture")]
 fn tze_outputs_hash(tze_outputs: &[TzeOut]) -> Blake2bHash {
     let mut data = vec![];
     for tzeout in tze_outputs {
@@ -201,6 +218,7 @@ pub enum SignableInput<'a> {
         script_code: &'a Script,
         value: Amount,
     },
+    #[cfg(feature = "zfuture")]
     Tze {
         index: usize,
         precondition: &'a Precondition,
@@ -217,6 +235,7 @@ impl<'a> SignableInput<'a> {
         }
     }
 
+    #[cfg(feature = "zfuture")]
     pub fn tze(index: usize, precondition: &'a Precondition, value: Amount) -> Self {
         SignableInput::Tze {
             index,
@@ -234,7 +253,14 @@ pub fn signature_hash_data<'a>(
 ) -> Vec<u8> {
     let sigversion = SigHashVersion::from_tx(tx);
     match sigversion {
-        SigHashVersion::Overwinter | SigHashVersion::Sapling | SigHashVersion::ZFuture => {
+        SigHashVersion::Sprout => unimplemented!(),
+
+        // ZIP-243 is implemented as a patch to ZIP-143; ZFuture is temporarily added
+        // as an additional patch to ZIP-243 but that match will likely be removed
+        // due to the need for transaction malleability fixes for TZE deployment.
+        //
+        // SigHashVersion::Overwinter | SigHashVersion::Sapling | SigHashVersion::ZFuture => {
+        _ => {
             let mut personal = [0; 16];
             (&mut personal[..12]).copy_from_slice(ZCASH_SIGHASH_PERSONALIZATION_PREFIX);
             (&mut personal[12..])
@@ -276,6 +302,7 @@ pub fn signature_hash_data<'a>(
             } else {
                 h.update(&[0; 32]);
             };
+            #[cfg(feature = "zfuture")]
             if sigversion == SigHashVersion::ZFuture {
                 update_hash!(
                     h,
@@ -293,7 +320,14 @@ pub fn signature_hash_data<'a>(
                 !tx.joinsplits.is_empty(),
                 joinsplits_hash(tx.version, &tx.joinsplits, &tx.joinsplit_pubkey.unwrap())
             );
-            if sigversion == SigHashVersion::Sapling || sigversion == SigHashVersion::ZFuture {
+            if sigversion == SigHashVersion::Sapling || {
+                #[cfg(feature = "zfuture")]
+                {
+                    sigversion == SigHashVersion::ZFuture
+                }
+                #[cfg(not(feature = "zfuture"))]
+                false
+            } {
                 update_hash!(
                     h,
                     !tx.shielded_spends.is_empty(),
@@ -307,7 +341,14 @@ pub fn signature_hash_data<'a>(
             }
             update_u32!(h, tx.lock_time, tmp);
             update_u32!(h, tx.expiry_height.into(), tmp);
-            if sigversion == SigHashVersion::Sapling || sigversion == SigHashVersion::ZFuture {
+            if sigversion == SigHashVersion::Sapling || {
+                #[cfg(feature = "zfuture")]
+                {
+                    sigversion == SigHashVersion::ZFuture
+                }
+                #[cfg(not(feature = "zfuture"))]
+                false
+            } {
                 h.update(&tx.value_balance.to_i64_le_bytes());
             }
             update_u32!(h, hash_type, tmp);
@@ -318,6 +359,7 @@ pub fn signature_hash_data<'a>(
                     script_code,
                     value,
                 } => {
+                    #[cfg(feature = "zfuture")]
                     let mut data = if sigversion == SigHashVersion::ZFuture {
                         // domain separation here is to avoid collision attacks
                         // between transparent and TZE inputs.
@@ -325,6 +367,9 @@ pub fn signature_hash_data<'a>(
                     } else {
                         vec![]
                     };
+
+                    #[cfg(not(feature = "zfuture"))]
+                    let mut data = vec![];
 
                     tx.vin[index].prevout.write(&mut data).unwrap();
                     script_code.write(&mut data).unwrap();
@@ -335,6 +380,7 @@ pub fn signature_hash_data<'a>(
                     h.update(&data);
                 }
 
+                #[cfg(feature = "zfuture")]
                 SignableInput::Tze {
                     index,
                     precondition,
@@ -353,6 +399,7 @@ pub fn signature_hash_data<'a>(
                     h.update(&data);
                 }
 
+                #[cfg(feature = "zfuture")]
                 SignableInput::Tze { .. } => {
                     panic!("A request has been made to sign a TZE input, but the signature hash version is not ZFuture");
                 }
@@ -362,7 +409,6 @@ pub fn signature_hash_data<'a>(
 
             h.finalize().as_ref().to_vec()
         }
-        SigHashVersion::Sprout => unimplemented!(),
     }
 }
