+++ zcash_primitives/src/transaction/tests.rs
@@ -1,133 +1,22 @@
 use ff::Field;
 use rand_core::OsRng;
 
-use proptest::collection::vec;
+#[cfg(all(feature = "test-dependencies", feature = "zfuture"))]
 use proptest::prelude::*;
-use proptest::sample::select;
 
-use crate::{
-    consensus::BranchId, constants::SPENDING_KEY_GENERATOR, extensions::transparent as tze,
-    legacy::Script, redjubjub::PrivateKey,
-};
+use crate::{constants::SPENDING_KEY_GENERATOR, redjubjub::PrivateKey};
+
+#[cfg(all(feature = "test-dependencies", feature = "zfuture"))]
+use crate::consensus::BranchId;
 
 use super::{
-    components::amount::MAX_MONEY,
-    components::{Amount, OutPoint, TxIn, TxOut, TzeIn, TzeOut},
+    components::Amount,
     sighash::{signature_hash, SignableInput},
-    Transaction, TransactionData, OVERWINTER_TX_VERSION, OVERWINTER_VERSION_GROUP_ID,
-    SAPLING_TX_VERSION, SAPLING_VERSION_GROUP_ID, ZFUTURE_TX_VERSION, ZFUTURE_VERSION_GROUP_ID,
+    Transaction, TransactionData,
 };
 
-prop_compose! {
-    fn arb_outpoint()(hash in prop::array::uniform32(1u8..), n in 1..(100 as u32)) -> OutPoint {
-        OutPoint::new(hash, n)
-    }
-}
-
-const VALID_OPCODES: [u8; 8] = [
-    0x00, // OP_FALSE,
-    0x51, // OP_1,
-    0x52, // OP_2,
-    0x53, // OP_3,
-    0xac, // OP_CHECKSIG,
-    0x63, // OP_IF,
-    0x65, // OP_VERIF,
-    0x6a, // OP_RETURN,
-];
-
-prop_compose! {
-    fn arb_script()(v in vec(select(&VALID_OPCODES[..]), 1..256)) -> Script {
-        Script(v)
-    }
-}
-
-prop_compose! {
-    fn arb_txin()(prevout in arb_outpoint(), script_sig in arb_script(), sequence in any::<u32>()) -> TxIn {
-        TxIn { prevout, script_sig, sequence }
-    }
-}
-
-prop_compose! {
-    fn arb_amount()(value in 0..MAX_MONEY) -> Amount {
-        Amount::from_i64(value).unwrap()
-    }
-}
-
-prop_compose! {
-    fn arb_txout()(value in arb_amount(), script_pubkey in arb_script()) -> TxOut {
-        TxOut { value, script_pubkey }
-    }
-}
-
-prop_compose! {
-    fn arb_witness()(extension_id in 0..(100 as u32), mode in (0..100 as u32), payload in vec(any::<u8>(), 32..256))  -> tze::Witness {
-        tze::Witness { extension_id, mode, payload }
-    }
-}
-
-prop_compose! {
-    fn arb_tzein()(prevout in arb_outpoint(), witness in arb_witness()) -> TzeIn {
-        TzeIn { prevout, witness }
-    }
-}
-
-prop_compose! {
-    fn arb_precondition()(extension_id in 0..(100 as u32), mode in (0..100 as u32), payload in vec(any::<u8>(), 32..256))  -> tze::Precondition {
-        tze::Precondition { extension_id, mode, payload }
-    }
-}
-
-prop_compose! {
-    fn arb_tzeout()(value in arb_amount(), precondition in arb_precondition()) -> TzeOut {
-        TzeOut { value, precondition }
-    }
-}
-
-fn tx_versions(branch_id: BranchId) -> impl Strategy<Value = (u32, u32)> {
-    match branch_id {
-        BranchId::Sprout => (1..(2 as u32)).prop_map(|i| (i, 0)).boxed(),
-        BranchId::Overwinter => Just((OVERWINTER_TX_VERSION, OVERWINTER_VERSION_GROUP_ID)).boxed(),
-        BranchId::ZFuture => Just((ZFUTURE_TX_VERSION, ZFUTURE_VERSION_GROUP_ID)).boxed(),
-        _otherwise => Just((SAPLING_TX_VERSION, SAPLING_VERSION_GROUP_ID)).boxed(),
-    }
-}
-
-prop_compose! {
-    fn arb_txdata(branch_id: BranchId)(
-        (version, version_group_id) in tx_versions(branch_id),
-        vin in vec(arb_txin(), 0..10),
-        vout in vec(arb_txout(), 0..10),
-        tze_inputs in vec(arb_tzein(), 0..10),
-        tze_outputs in vec(arb_tzeout(), 0..10),
-        lock_time in any::<u32>(),
-        expiry_height in any::<u32>(),
-        value_balance in arb_amount(),
-    ) -> TransactionData {
-        TransactionData {
-            overwintered: branch_id != BranchId::Sprout,
-            version,
-            version_group_id,
-            vin, vout,
-            tze_inputs:  if branch_id == BranchId::ZFuture { tze_inputs } else { vec![] },
-            tze_outputs: if branch_id == BranchId::ZFuture { tze_outputs } else { vec![] },
-            lock_time,
-            expiry_height: expiry_height.into(),
-            value_balance,
-            shielded_spends: vec![], //FIXME
-            shielded_outputs: vec![], //FIXME
-            joinsplits: vec![], //FIXME
-            joinsplit_pubkey: None, //FIXME
-            joinsplit_sig: None, //FIXME
-            binding_sig: None, //FIXME
-        }
-    }
-}
-
-prop_compose! {
-    fn arb_tx(branch_id: BranchId)(tx_data in arb_txdata(branch_id)) -> Transaction {
-        Transaction::from_data(tx_data).unwrap()
-    }
-}
+#[cfg(all(feature = "test-dependencies", feature = "zfuture"))]
+use super::testing::arb_tx;
 
 #[test]
 fn tx_read_write() {
@@ -186,6 +75,7 @@ fn tx_write_rejects_unexpected_binding_sig() {
     }
 }
 
+#[cfg(all(feature = "zfuture", feature = "test-dependencies"))]
 proptest! {
     #[test]
     fn test_tze_roundtrip(tx in arb_tx(BranchId::ZFuture)) {
@@ -207,6 +97,7 @@ proptest! {
 }
 
 #[test]
+#[cfg(feature = "zfuture")]
 fn test_tze_tx_parse() {
     let txn_bytes = vec![
         0xFF, 0xFF, 0x00, 0x80, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x01, 0x52, 0x52, 0x52, 0x52,
